# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Mood = (typeof Mood)[keyof typeof Mood];

export const Mood = {
  great: "great",
  good: "good",
  okay: "okay",
  bad: "bad",
  awful: "awful",
} as const;

export interface Category {
  id: string;
  name: string;
  color: string;
  createdAt: string;
}

export type ListCategoriesResponse = Category[];

export interface CreateCategoryBody {
  name: string;
  color: string;
}

export interface UpdateCategoryBody {
  name?: string;
  color?: string;
}

export interface Entry {
  id: string;
  text: string;
  mood: Mood;
  categoryIds: string[];
  createdAt: string;
  updatedAt: string;
}

export type ListEntriesResponse = Entry[];

export interface CreateEntryBody {
  text: string;
  mood: Mood;
  categoryIds: string[];
}

export interface UpdateEntryBody {
  text?: string;
  mood?: Mood;
  categoryIds?: string[];
}

export interface MoodCount {
  mood: Mood;
  count: number;
}

export interface CategoryCount {
  categoryId: string;
  categoryName: string;
  count: number;
}

export interface EntryStats {
  totalEntries: number;
  currentStreak: number;
  moodCounts: MoodCount[];
  categoryCounts: CategoryCount[];
}

export type ListEntriesParams = {
  /**
   * Full-text search across entry text
   */
  search?: string;
  /**
   * Filter entries by category ID
   */
  categoryId?: string;
};