# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns entries ordered by created_at desc, with optional search and category filter
 * @summary List all entries
 */
export const ListEntriesQueryParams = zod.object({
  search: zod.coerce
    .string()
    .optional()
    .describe("Full-text search across entry text"),
  categoryId: zod.coerce
    .string()
    .optional()
    .describe("Filter entries by category ID"),
});

export const ListEntriesResponseItem = zod.object({
  id: zod.string(),
  text: zod.string(),
  mood: zod.enum(["great", "good", "okay", "bad", "awful"]),
  categoryIds: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListEntriesResponse = zod.array(ListEntriesResponseItem);

/**
 * @summary Create a new entry
 */
export const CreateEntryBody = zod.object({
  text: zod.string(),
  mood: zod.enum(["great", "good", "okay", "bad", "awful"]),
  categoryIds: zod.array(zod.string()),
});

/**
 * Returns counts by mood, by category, recent streak, and total count
 * @summary Get diary stats
 */
export const GetEntryStatsResponse = zod.object({
  totalEntries: zod.number(),
  currentStreak: zod.number(),
  moodCounts: zod.array(
    zod.object({
      mood: zod.enum(["great", "good", "okay", "bad", "awful"]),
      count: zod.number(),
    }),
  ),
  categoryCounts: zod.array(
    zod.object({
      categoryId: zod.string(),
      categoryName: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get a single entry
 */
export const GetEntryParams = zod.object({
  id: zod.coerce.string(),
});

export const GetEntryResponse = zod.object({
  id: zod.string(),
  text: zod.string(),
  mood: zod.enum(["great", "good", "okay", "bad", "awful"]),
  categoryIds: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update an entry
 */
export const UpdateEntryParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateEntryBody = zod.object({
  text: zod.string().optional(),
  mood: zod.enum(["great", "good", "okay", "bad", "awful"]).optional(),
  categoryIds: zod.array(zod.string()).optional(),
});

export const UpdateEntryResponse = zod.object({
  id: zod.string(),
  text: zod.string(),
  mood: zod.enum(["great", "good", "okay", "bad", "awful"]),
  categoryIds: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an entry
 */
export const DeleteEntryParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  color: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Create a new category
 */
export const CreateCategoryBody = zod.object({
  name: zod.string(),
  color: zod.string(),
});

/**
 * @summary Update a category name or color
 */
export const UpdateCategoryParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateCategoryBody = zod.object({
  name: zod.string().optional(),
  color: zod.string().optional(),
});

export const UpdateCategoryResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  color: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a category
 */
export const DeleteCategoryParams = zod.object({
  id: zod.coerce.string(),
});